refactor: move redis to database package
