+++ src/main/java/de/rexlmanu/fairychat/plugin/core/playerchat/PlayerChatMessageSubscriber.java
@@ -4,7 +4,7 @@ import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import de.rexlmanu.fairychat.plugin.Constants;
 import de.rexlmanu.fairychat.plugin.configuration.FairyChatConfiguration;
-import de.rexlmanu.fairychat.plugin.redis.channel.RedisChannelSubscriber;
+import de.rexlmanu.fairychat.plugin.database.redis.channel.RedisChannelSubscriber;
 import lombok.RequiredArgsConstructor;
 import org.bukkit.Server;
