+++ src/main/java/de/rexlmanu/fairychat/plugin/core/user/redis/channel/RedisUserLogoutSubscriber.java
@@ -3,7 +3,7 @@ package de.rexlmanu.fairychat.plugin.core.user.redis.channel;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import de.rexlmanu.fairychat.plugin.core.user.redis.RedisUserService;
-import de.rexlmanu.fairychat.plugin.redis.channel.RedisChannelSubscriber;
+import de.rexlmanu.fairychat.plugin.database.redis.channel.RedisChannelSubscriber;
 import lombok.RequiredArgsConstructor;
 
 @Singleton
